优化网格策略

+++ strategy/strategy.go
@@ -1,7 +1,8 @@
 package strategy
 
+// 策略模块，传入价格，返回买卖动作
 type St interface {
 	GetName() string
-	SendPrice(f float32)
+	SendPrice(f float64)
 	Close()
 }
